ai-sketch-completer/requirements.txt
# AI Sketch Completer - Requirements
# Python 3.8+ recommended

# Core dependencies
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0

# Image processing
Pillow>=10.0.0
numpy>=1.24.0

# Optional: For faster model loading
safetensors>=0.3.0

# Note: For CUDA support, install PyTorch with CUDA:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
